+++ app/controllers/services_controller.rb
@@ -13,4 +13,13 @@ class ServicesController < ApplicationController
     @service = Service.find(params[:id])
     @service_opinions = ServiceOpinion.joins(:project_item).where(project_items: { service: @service })
   end
+
+  def active?(service)
+    # user can have only one open_access service order with status ready
+    service.open_access &&
+    current_user &&
+    ProjectItem.find_by(service: service.id,
+                                               project: current_user.projects,
+                                         status: [:created, :ready])
+  end
 end