+++ app/helpers/service_helper.rb
@@ -18,7 +18,7 @@ module ServiceHelper
     result.html_safe
   end
 
-  def write_button_text
+  def project_item_service_text
     @service.open_access ? "Add to my services" : "Order"
   end