+++ app/policies/project_item/question_policy.rb
@@ -2,7 +2,7 @@
 
 class ProjectItem::QuestionPolicy < ApplicationPolicy
   def create?
-    record.project_item.active? && record.project_item.user == user
+    record.project_item.draft? && record.project_item.user == user
   end
 
   def permitted_attributes