+++ spec/factories/project_items.rb
@@ -5,5 +5,9 @@ FactoryBot.define do
     status :created
     service
     project
+
+    factory :open_access_project_item do
+      status :ready
+    end
   end
 end