+++ spec/policies/project_item_policy_spec.rb
@@ -13,7 +13,7 @@ RSpec.describe ProjectItemPolicy do
     end
   end
 
-  permissions :create? do
+  permissions :create?, :update? do
     it "grants access to create item in owned project" do
       expect(subject).
         to permit(user, build(:project_item,